requirements.txt
#ipykernel
#ipywidgets
#PyMuPDF
#chromadb
#sentence-transformers
python-dotenv
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters
Flask
gunicorn # heroku
openai
llama-index
weaviate-client
llama-index-vector-stores-weaviate
Babel
google-cloud-translate